Extract UnavailableWorkPackage to its own file
